refactor: split App.tsx
