+++ playground/src/main.tsx
@@ -1,6 +1,5 @@
 import { StrictMode } from "react";
 import { createRoot } from "react-dom/client";
-import { ThemeProvider } from "./contexts";
 import "./styles/index.css";
 import App from "./App";
 import { initMonaco } from "./utils/monacoThemes";
@@ -9,8 +8,6 @@ await initMonaco();
 
 createRoot(document.getElementById("root")!).render(
   <StrictMode>
-    <ThemeProvider>
-      <App />
-    </ThemeProvider>
+    <App />
   </StrictMode>,
 );
